Moved release related tasks and plugins to appropriate package
