+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.publications-comparator.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.PublicationsComparatorPlugin
+implementation-class=org.shipkit.internal.gradle.release.PublicationsComparatorPlugin
