+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.release-needed.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.ReleaseNeededPlugin
+implementation-class=org.shipkit.internal.gradle.release.ReleaseNeededPlugin
