+++ src/test/groovy/org/shipkit/internal/gradle/release/ReleaseNeededPluginTest.groovy
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle
+package org.shipkit.internal.gradle.release
 
 import testutil.PluginSpecification
